report of Elearning community website
eLearning-Community is an open-source platform website, which provide 
solutions to complex coding problems in several languages and provide 
free tutorials for beginners, and it is built using more than 1500 lines of 
codes of HTML, CSS and a little bit of JavaScript. The main attraction of 
the website is that you can get free codes of different languages such as 
Java, C and python etc at one place. 
There are several other pages also provided to facilitate the users for exLogin page, contact page, about page and sign up page .